requirements.txt
numpy
Pillow
cmake
dlib
face_recognition